Atlas-web: some small bug fixes to restrict the scope of NetCDF data returned better.  Migrating some more code to new package names.  Fixed some horrid find/replace problem that broke several jsp files.  Modified atlas-web version to fall into line with others - we need to decide version number strategy before formal full release.
Also, put in checks for AtlasRFactory in application listener, so tomcat won't start webapp if R is not configured correctly.  This is not perfect - the underlying jri library still throws errors if R is badly setup, but does give an extra safeguard.  There are still issues to work out though.

git-svn-id: svn://bar.ebi.ac.uk/branches/atlas-standalone@9881 2913f559-6b04-0410-9a09-c530ee9f5186